visibility of the follow/unfollow buttons changes based on follow/unfollow
