+++ src/js/api/profile/followService.js
@@ -1,8 +1,10 @@
-import { loadUserProfileAndPosts } from "../../router/views/profile.js";
+// import { loadUserProfileAndPosts } from "../../router/views/profile.js";
 import { showErrorAlert } from "../../ui/global/alertHandler.js";
+import { updateFollowButtons } from "../../ui/profile/followBtns.js";
 import { API_SOCIAL_PROFILES } from "../constants.js";
 import { headers } from "../headers.js";
 
+// Follow a user
 export async function followUser(username) {
   try {
     const response = await fetch(`${API_SOCIAL_PROFILES}/${username}/follow`, {
@@ -15,13 +17,16 @@ export async function followUser(username) {
     }
 
     const data = await response.json();
-    console.log("Followed users", data);
-    loadUserProfileAndPosts(username);
+    console.log("Followed user:", data);
+
+    // Update UI to show that the user is followed
+    updateFollowButtons(username, true);
   } catch (error) {
     showErrorAlert(`Error following user: ${error.message}`);
   }
 }
 
+// Unfollow a user
 export async function unfollowUser(username) {
   try {
     const response = await fetch(
@@ -38,7 +43,9 @@ export async function unfollowUser(username) {
 
     const data = await response.json();
     console.log("Unfollowed user:", data);
-    loadUserProfileAndPosts(username);
+
+    // Update UI to show that the user is unfollowed
+    updateFollowButtons(username, false);
   } catch (error) {
     showErrorAlert(`Error unfollowing user: ${error.message}`);
   }
