--- src/js/ui/profile/follow.js
@@ -1,14 +1,0 @@
-import { followUser, unfollowUser } from "../../api/profile/followService";
-
-export function followButtonsListener(username) {
-  const followButton = document.getElementById("follow-btn");
-  const unfollowButton = document.getElementById("unfollow-btn");
-
-  followButton.addEventListener("click", () => {
-    followUser(username);
-  });
-
-  unfollowButton.addEventListener("click", () => {
-    unfollowUser(username);
-  });
-}
